+++ CHANGELOG.rst
@@ -26,6 +26,8 @@ Added
 
 - New GraphQL API ``failover_state_provider_status`` to ping state provider connection.
 
+- New issue about unhealthy replicasets.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~